JobSearcher/Utils/VisibilityToBooleanConverter.cs
using System;
using System.Globalization;
using System.Windows;
using System.Windows.Data;
using System.Windows.Markup;

namespace JobSearcher.Utils
{
    [MarkupExtensionReturnType(typeof(IValueConverter))]
    public class VisibilityToBooleanConverter : MarkupExtension, IValueConverter
    {
        #region Public Methods

        public override object ProvideValue(IServiceProvider serviceProvider)
        {
            return this;
        }

        #endregion

        #region IValueConverter Members

        public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
        {
            return value != null && (Visibility)value == Visibility.Visible;
        }

        public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
        {
            return value != null && (bool)value ? Visibility.Visible : Visibility.Collapsed;
        }

        #endregion
    }
}